Make gh-emoji optional
gh-emoji crate includes *images* of GitHub's emoji - this is quite a big
dependency. It increases the binary size by 1 MiB; that's +25 % when
building v0.18.0 on Alpine Linux with build flags to optimize size.
I consider it an unnecessary bloat that should be optional.
